move utilities to Util
